Modify: createTempFeed API, upload API  - createTempFeed API   - 임시저장 실행시, 해당 사용자의 uploadfile중 feedID가 없는 entity 삭제 및 s3 파일 역시 삭제   - transaction 처리 (실패시 모두 rollback)
 - upload API
  - createTempFeed API에서 upload.service.ts에서의 deleteUploadFile 함수 재사용시 transaction 및 Entity 삭제 중복으로 인한 response delay 문제 해결

+++ src/services/upload.service.ts
@@ -92,7 +92,12 @@ const deleteUploadFile = async (file_links: string[]): Promise<void> => {
   // mySQL에서 file_link를 통해 uploadFile의 ID를 담을 배열
   let uploadFileIdArray = [];
 
-  for (const file_link of file_links) {
+  // feeds.service에서 본 함수를 사용할때, mySQL의 테이블에서는 삭제할 필요가 없기때문에 조건을 만들어준다.
+  const newFileLinks = file_links.filter(
+    (file_link: string) => file_link !== 'deleteUploadFile.constroller'
+  );
+
+  for (const file_link of newFileLinks) {
     const findFile = await dataSource.manager.findOneOrFail<UploadFiles>(
       UploadFiles,
       {
@@ -133,8 +138,10 @@ const deleteUploadFile = async (file_links: string[]): Promise<void> => {
     }
   }
 
-  // mySQL에서 개체 삭제
-  await dataSource.manager.delete(UploadFiles, uploadFileIdArray);
+  if (file_links.includes('deleteUploadFile.constroller')) {
+    // mySQL에서 개체 삭제
+    await dataSource.manager.delete(UploadFiles, uploadFileIdArray);
+  }
 };
 
 export default { uploadFiles, deleteUploadFile };
